Per file ACLs: Enable in the fileserver
This patch enables per file ACL support in the fileserver.

Existing StoreACL and FetchACL RPCs are allowed to deal with
file ACLs.  Access rights are based are based on a file ACL if
one exists.
Some semantics that will evolve in later patches:
- Setting a directory ACL never affects the children - we'd prefer
to at least affect files that have never had an ACL set
- New ACL RPCs will add per-file semantics.  The existing RPCs will
revert to a legacy behaviour

Change-Id: I889c34ccd3fad33809f8df44a52d237bfa008f10