Per-file ACLs: Dump and volume move support
Support the inclusion of file ACLs in dumps and clones, allowing
the ACLs to survive dump/restore and volume moves.

Add new dump tag for file ACLs.  For each file that has an ACL,
fetch the ACL data from the ACL file and include in the dump.  On
restore, allocate slots as needed in the new volume.

Adjust cloning code to copy ACL data as needed to the clone.

With this commit, vos dump/restore, vos move and vos release of
replicated volumes is now supported with file ACLs.

Remaining tasks:
- Fixing the standalone restore and dump applications
- Adjusting the dump size estimates to account for ACLs

Change-Id: Iee7a59e3abf7a5b7aa17000b09e41e353197e2c1