Per-file ACLs: minor cleanup in vnode load/store
Cleanup mostly comments and code formatting.
Need a better way to deal with legacy vnode magic values.  When
we notice on load, we might not have a write lock to modify the
vnode on store.
Of course if we eventually simply have a field in an extended
vnode, this becomes a non-issue.