Per-file ACLs: Introduce new ACL RPCs
Introduce new versions of the FetchACL and StoreACL RPC calls.
These are identical to the originals except that they are defined
to allow storing and fetching file ACLs.
- StoreACL2 on a file is allowed, stores a file ACL
- FetchACL2 returns the file's ACL if there is one, the parent
otherwise
- FetchACL2 should eventually return an error code for a file with
no ACL, instead of the parent ACL.  The client code can then
be aware of the absence of a file ACL.

Old RPCs return to their original behaviour.

SetAccessList now checks the client's capability flags to determine
if it is file ACL aware.