rx: Experimental hard ack change
As an experiment, set the hard ack quota to be half the window size.
I suspect we actually need to be clever about this - we only need to
send hard Acks when we're not going to be sending a soft ack shortly
(ie, when the peer's transmit window is full, and they need the window
to move before they can send any more to us)

However, knowing where the peer's window is is hard. We know what we've
told them in terms of our receive window, but we don't know what
limitations they're working under in terms of congestion and soft start,
which will limit their window size.

Change-Id: I3bbba825ecd1c494a2be67c18820bedd5fccfdef