unix cm  rx-oblivious connection pooling
Implements a connection pool for Rx client connections with the
same credentials.  The code trivially avoids the limit on Rx
call channels, without touching Rx directly.  The conn call limit
is known to be hit in cache-bypass, and probably other use cases,
so there is an incentive to address it potentially sooner than
larger Rx changes are ready to merge.

Change-Id: I956aa22cd52b1c43187c7e03230e36820ba11c4d