CM connection pooling: Make the connection vector size tunable.
Add a new pioctl and the corresponding fs command to set or read
the currrent maximum number of concurrent RX connections per
security context.  The allowed range is from 1 to 10, and the
data structures are statically allocated to the largest possible
size.

Syntax: fs maxconns <N>

Examples:

Read current maximum:
  fs maxconns
Set maximum to 4:
  fs maxconns 4
  fs maxconns -conns 4

Change-Id: Ia6c2b060707f52cb4d1d59e7970394ad57780495