ubik: add DISK_GetFileV2
Add the new version of the GetFile RPC.  This has a different
signature and behaviour than the old version, so the old one is
left as is.

The file size sent by the RPC is a 64-bit value.  haveVersion is
an indication by the caller of the version he already has - if it
matches the current version, send no data and simply return success.

Change-Id: I43f0f2ded461755b4f9645cb76db8e54261c706b