RW Replication: basic bits
This adds the client side code to ask for the creation of a new
RW replica for a volume.  This is done with the addition of a new
-rw flag to vos addsite, which will add a new site in the VLDB
with the new RWSLAVEVOL flag introduced here.

Adjust output of vos examine and vos listvldb.

There is no server code at this point to deal with this flag.

While we're modifying the UV_AddSite function, rename the current
one from AddSite2 to AddSite, and remove the unused old version
which has no callers.

Change-Id: I3aab49b3b5265adab494479041f1317659f9a970