RW Replication: client changes
Some changes to the client to cope with RW slave volumes:
- Add a new rwserver field in the volume structure to keep track
of the RW server
- When looking for the RW volume, include any RW slave servers in
the list
- Add a new parameter to afs_Conn to force the connection to the
RW server.  Adjust all callers to set this flag appropriately,
depending on the operation

The general idea is that all operations that modify the volume
go to the single RW server site, while other operations go to
the lowest ranked site, RW or RW slave.

Change-Id: Id72cf41679ff804271eb60e2ad7b2c04099dfeb0