RW replication: Stick to RW server after calling MakeDir
Add the ability to tell afs_GetVCache and afs_FetchStatus to use
the RW server only.  This is needed for example when the client
issues a FetchStatus right after a MakeDir RPC.  If the FetchStatus
is directed to a RW replica, it might reach the server before the
replica gets the remote MakeDir call from the RW server, resulting
in a return of ENOENT to mkdir.

Change-Id: Ic17e4cfd20b804cff7d21864c0269500f03b24a1