RW replication: Insert RemoveFile op into list before releasing locks
In order to insure that updates are inserted into the queue in the
correct order, we need to do so before releasing the lock on the
vnode in PutVolumePackage.  Otherwise, a later update touching the
same vnode might get inserted before us and sent to the slaves in
the wrong order.  The oberved case was a RemoveFile on a Vnode
followed by a CreateFile that is given the same Vnode.

A similar change will be needed for some other operations.

Change-Id: I43c0f89d68dc73e43de8cfe08ab30b15f2c84913