RW Replication: Add and remove RW replication sites
Add the ability to specify and remove RW replication sites for
a volume.

A new -readwrite option is added to vos addsite.  When specified,
the volume at the new site will be flagged as a read/write replica
in the VLDB.  Description of the new option is added to the man
page.

Change-Id: I487c8002b07ad614d874fa6e72b57ca181b0b55e