Define new filserver to fileserver replication interface
Add a new fileserver to fileserver interface for read/write
replication.

This interface will be used to forward updates from the fileserver
holding the master volume to the fileservers holding the replicas.

The calls are adpated versions of the RXAFS calls that modify
volume data:
- All OUT parameters are removed
- A new parameter passes the ID of the client making the change
- For calls that create vnodes, a parameter is added to pass in
the vnode number to be created

Change-Id: I2d42fa137ce91a19ad8b5ee3b3d079605766e9f2