viced: Add support for StoreData replication
The code is restructured a bit to make a new SAFS_StoreData64 that
can be called by the remote StoreData RPC.

StoreData_RXStyle, if provided with a non NULL buffer pointer, will
allocate space for the written data and make a copy of it in this
buffer.  When replication is active for a volume, this will be used
to get a buffer to replay the StoreData to the remote replicas.

Change-Id: Ib47e17b0e0bd6ca9433a53fad05ed11cfa6ac87c